requirements.txt
anthropic>=0.40.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
streamlit>=1.40.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
